aura: remove unneeded SlotDuration struct and rename digest -> digests
